requirements.txt


dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord.py==1.2.5
python-dotenv==0.10.2
requests==2.7.0
bs4==0.0.1
googletrans==2.4.0
requests-html==0.10.0
praw==6.4.0 
twython==3.7.0

pandas==0.25.3 
psycopg2-binary==2.8.4
sqlalchemy==1.3.13 
nltk==3.4.5 
numpy==1.16.1
geopy==1.20.0
gmplot==1.2.0 
folium==0.10.1
setuptools==40.6.3

git+https://github.com/AdilZouitine/pyFeel 
